+++ instrumentation/opentelemetry-instrumentation-dbapi/tests/test_dbapi_integration.py
@@ -14,6 +14,7 @@
 
 # pylint: disable=too-many-lines
 
+import asyncio
 import logging
 import re
 from unittest import mock
@@ -21,6 +22,10 @@ from unittest import mock
 from opentelemetry import context
 from opentelemetry import trace as trace_api
 from opentelemetry.instrumentation import dbapi
+from opentelemetry.instrumentation._semconv import (
+    OTEL_SEMCONV_STABILITY_OPT_IN,
+    _OpenTelemetrySemanticConventionStability,
+)
 from opentelemetry.instrumentation.utils import suppress_instrumentation
 from opentelemetry.sdk import resources
 from opentelemetry.semconv._incubating.attributes import net_attributes
@@ -34,6 +39,20 @@ from opentelemetry.semconv._incubating.attributes.net_attributes import (
     NET_PEER_NAME,
     NET_PEER_PORT,
 )
+from opentelemetry.semconv._incubating.metrics.db_metrics import (
+    DB_CLIENT_OPERATION_DURATION,
+    DB_CLIENT_RESPONSE_RETURNED_ROWS,
+)
+from opentelemetry.semconv.attributes.db_attributes import (
+    DB_NAMESPACE,
+    DB_OPERATION_NAME,
+    DB_SYSTEM_NAME,
+)
+from opentelemetry.semconv.attributes.error_attributes import ERROR_TYPE
+from opentelemetry.semconv.attributes.server_attributes import (
+    SERVER_ADDRESS,
+    SERVER_PORT,
+)
 from opentelemetry.test.test_base import TestBase
 
 
@@ -41,6 +60,8 @@ from opentelemetry.test.test_base import TestBase
 class TestDBApiIntegration(TestBase):
     def setUp(self):
         super().setUp()
+        # Reset cached stability opt-in so each test re-reads os.environ.
+        _OpenTelemetrySemanticConventionStability._initialized = False
         self.tracer = self.tracer_provider.get_tracer(__name__)
 
     def test_span_succeeded(self):
@@ -252,6 +273,185 @@ class TestDBApiIntegration(TestBase):
 
         spans_list = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(spans_list), 0)
+
+    def _get_metric(self, name):
+        return next(
+            (
+                metric
+                for metric in self.get_sorted_metrics()
+                if metric.name == name
+            ),
+            None,
+        )
+
+    def test_metrics_not_emitted_in_default_mode(self):
+        # Without OTEL_SEMCONV_STABILITY_OPT_IN=database, no metrics are exported
+        db_integration = dbapi.DatabaseApiIntegration(
+            "instrumenting_module_test_name", "testcomponent"
+        )
+        mock_connection = db_integration.wrapped_connection(
+            mock_connect, (), {}
+        )
+        cursor = mock_connection.cursor()
+        cursor.execute("SELECT 1", rowcount=3)
+
+        self.assertIsNone(self._get_metric(DB_CLIENT_OPERATION_DURATION))
+        self.assertIsNone(self._get_metric(DB_CLIENT_RESPONSE_RETURNED_ROWS))
+
+    @mock.patch.dict("os.environ", {OTEL_SEMCONV_STABILITY_OPT_IN: "database"})
+    def test_operation_duration_recorded(self):
+        connection_props = {
+            "database": "testdatabase",
+            "server_host": "testhost",
+            "server_port": 123,
+            "user": "testuser",
+        }
+        connection_attributes = {
+            "database": "database",
+            "port": "server_port",
+            "host": "server_host",
+            "user": "user",
+        }
+        db_integration = dbapi.DatabaseApiIntegration(
+            "instrumenting_module_test_name",
+            "testcomponent",
+            connection_attributes,
+        )
+        mock_connection = db_integration.wrapped_connection(
+            mock_connect, (), connection_props
+        )
+        cursor = mock_connection.cursor()
+        cursor.execute("SELECT * FROM users")
+
+        duration_metric = self._get_metric(DB_CLIENT_OPERATION_DURATION)
+        self.assertIsNotNone(duration_metric)
+        self.assertEqual(duration_metric.unit, "s")
+        points = list(duration_metric.data.data_points)
+        self.assertEqual(len(points), 1)
+        attributes = dict(points[0].attributes)
+        self.assertEqual(attributes[DB_SYSTEM_NAME], "testcomponent")
+        self.assertEqual(attributes[DB_NAMESPACE], "testdatabase")
+        self.assertEqual(attributes[DB_OPERATION_NAME], "SELECT")
+        self.assertEqual(attributes[SERVER_ADDRESS], "testhost")
+        self.assertEqual(attributes[SERVER_PORT], 123)
+        self.assertNotIn(ERROR_TYPE, attributes)
+        self.assertEqual(points[0].count, 1)
+        self.assertGreaterEqual(points[0].sum, 0.0)
+
+    @mock.patch.dict("os.environ", {OTEL_SEMCONV_STABILITY_OPT_IN: "database"})
+    def test_operation_duration_error_type(self):
+        db_integration = dbapi.DatabaseApiIntegration(
+            "instrumenting_module_test_name", "testcomponent"
+        )
+        mock_connection = db_integration.wrapped_connection(
+            mock_connect, (), {}
+        )
+        cursor = mock_connection.cursor()
+        with self.assertRaises(Exception):
+            cursor.execute("SELECT 1", throw_exception=True)
+
+        duration_metric = self._get_metric(DB_CLIENT_OPERATION_DURATION)
+        self.assertIsNotNone(duration_metric)
+        points = list(duration_metric.data.data_points)
+        self.assertEqual(len(points), 1)
+        attributes = dict(points[0].attributes)
+        self.assertEqual(attributes[ERROR_TYPE], "Exception")
+        self.assertEqual(attributes[DB_OPERATION_NAME], "SELECT")
+
+        # returned_rows should NOT be recorded on error
+        rows_metric = self._get_metric(DB_CLIENT_RESPONSE_RETURNED_ROWS)
+        self.assertIsNone(rows_metric)
+
+    @mock.patch.dict("os.environ", {OTEL_SEMCONV_STABILITY_OPT_IN: "database"})
+    def test_returned_rows_recorded(self):
+        db_integration = dbapi.DatabaseApiIntegration(
+            "instrumenting_module_test_name", "testcomponent"
+        )
+        mock_connection = db_integration.wrapped_connection(
+            mock_connect, (), {}
+        )
+        cursor = mock_connection.cursor()
+        cursor.execute("SELECT 1", rowcount=3)
+
+        rows_metric = self._get_metric(DB_CLIENT_RESPONSE_RETURNED_ROWS)
+        self.assertIsNotNone(rows_metric)
+        self.assertEqual(rows_metric.unit, "{row}")
+        points = list(rows_metric.data.data_points)
+        self.assertEqual(len(points), 1)
+        self.assertEqual(points[0].sum, 3)
+        self.assertEqual(points[0].count, 1)
+
+    @mock.patch.dict("os.environ", {OTEL_SEMCONV_STABILITY_OPT_IN: "database"})
+    def test_returned_rows_skipped_when_rowcount_unknown(self):
+        db_integration = dbapi.DatabaseApiIntegration(
+            "instrumenting_module_test_name", "testcomponent"
+        )
+        mock_connection = db_integration.wrapped_connection(
+            mock_connect, (), {}
+        )
+        cursor = mock_connection.cursor()
+        # default rowcount is -1 on MockCursor
+        cursor.execute("CREATE TABLE t (x INT)")
+
+        rows_metric = self._get_metric(DB_CLIENT_RESPONSE_RETURNED_ROWS)
+        self.assertIsNone(rows_metric)
+        # duration is still recorded
+        self.assertIsNotNone(self._get_metric(DB_CLIENT_OPERATION_DURATION))
+
+    @mock.patch.dict("os.environ", {OTEL_SEMCONV_STABILITY_OPT_IN: "database"})
+    def test_custom_meter_provider(self):
+        meter_provider, metrics_reader = self.create_meter_provider()
+        db_integration = dbapi.DatabaseApiIntegration(
+            "instrumenting_module_test_name",
+            "testcomponent",
+            meter_provider=meter_provider,
+        )
+        mock_connection = db_integration.wrapped_connection(
+            mock_connect, (), {}
+        )
+        cursor = mock_connection.cursor()
+        cursor.execute("SELECT 1", rowcount=1)
+
+        metrics_data = metrics_reader.get_metrics_data()
+        names = {
+            m.name
+            for rm in metrics_data.resource_metrics
+            for sm in rm.scope_metrics
+            for m in sm.metrics
+        }
+        self.assertIn(DB_CLIENT_OPERATION_DURATION, names)
+        self.assertIn(DB_CLIENT_RESPONSE_RETURNED_ROWS, names)
+
+    @mock.patch.dict("os.environ", {OTEL_SEMCONV_STABILITY_OPT_IN: "database"})
+    def test_async_operation_duration_recorded(self):
+        db_integration = dbapi.DatabaseApiIntegration(
+            "instrumenting_module_test_name", "testcomponent"
+        )
+        cursor_tracer = dbapi.CursorTracer(db_integration)
+        mock_cursor = MockCursor()
+
+        async def async_execute(_query, rowcount=-1):
+            mock_cursor.rowcount = rowcount
+            return None
+
+        asyncio.run(
+            cursor_tracer.traced_execution_async(
+                mock_cursor, async_execute, "SELECT 1", rowcount=5
+            )
+        )
+
+        duration_metric = self._get_metric(DB_CLIENT_OPERATION_DURATION)
+        self.assertIsNotNone(duration_metric)
+        points = list(duration_metric.data.data_points)
+        self.assertEqual(len(points), 1)
+        self.assertEqual(
+            dict(points[0].attributes)[DB_OPERATION_NAME], "SELECT"
+        )
+
+        rows_metric = self._get_metric(DB_CLIENT_RESPONSE_RETURNED_ROWS)
+        self.assertIsNotNone(rows_metric)
+        rows_points = list(rows_metric.data.data_points)
+        self.assertEqual(rows_points[0].sum, 5)
 
     def test_commenter_options_propagation(self):
         db_integration = dbapi.DatabaseApiIntegration(
@@ -1304,6 +1504,7 @@ class MockCursor:
         self.query = ""
         self.params = None
         self.connection = None
+        self.rowcount = -1
         # Mock mysql.connector modules and method
         self._cnx = mock.MagicMock()
         self._cnx._cmysql = mock.MagicMock()
@@ -1313,24 +1514,29 @@ class MockCursor:
         self._items = []
 
     # pylint: disable=unused-argument, no-self-use
-    def execute(self, query, params=None, throw_exception=False):
+    def execute(self, query, params=None, throw_exception=False, rowcount=-1):
         if throw_exception:
             # pylint: disable=broad-exception-raised
             raise Exception("Test Exception")
+        self.rowcount = rowcount
 
     def __iter__(self):
         yield from self._items
 
     # pylint: disable=unused-argument, no-self-use
-    def executemany(self, query, params=None, throw_exception=False):
+    def executemany(
+        self, query, params=None, throw_exception=False, rowcount=-1
+    ):
         if throw_exception:
             # pylint: disable=broad-exception-raised
             raise Exception("Test Exception")
         self.query = query
         self.params = params
+        self.rowcount = rowcount
 
     # pylint: disable=unused-argument, no-self-use
-    def callproc(self, query, params=None, throw_exception=False):
+    def callproc(self, query, params=None, throw_exception=False, rowcount=-1):
         if throw_exception:
             # pylint: disable=broad-exception-raised
             raise Exception("Test Exception")
+        self.rowcount = rowcount